v2.02 [BUGFIX] apacman crash
Recent changes to apacman exposed a bug (in apacman) that causes it to
crash when run in the background. See issue 2 for more information.

Ideally, the authors of this tool would fix this in apacman. Should
they do this in the future, this workaround can be removed.

The current fix is a workaround forcing apacman to use a dummy stty
command that will always return a numerical value, thus avoiding the
crash.

Also in this commit are a couple minor code-cleanup changes.

Signed-off-by: Lectrode <electrodexsnet@gmail.com>